final-pjt-back/requirement.txt
asgiref==3.5.1
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==3.2.12
django-allauth==0.50.0
django-bootstrap-v5==1.0.5
django-cors-headers==3.12.0
django-extensions==3.1.5
django-pandas==0.6.6
django-seed==0.3.1
djangorestframework==3.13.1
drf-yasg==1.20.0
executing==0.8.3
Faker==13.3.4
idna==3.3
importlib-metadata==3.10.1
inflection==0.5.1
install==1.3.5
ipython==8.2.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.29
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
PyJWT==2.4.0
pyparsing==3.0.8
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
sqlparse==0.4.2
stack-data==0.2.0
threadpoolctl==3.1.0
toposort==1.7
traitlets==5.1.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
zipp==3.8.0
